Pertemuan 6 - StudentController

+++ rest-api/app/Http/Controllers/StudentController.php
@@ -8,25 +8,40 @@ use App\Models\Student;
 
 class StudentController extends Controller
 {
+    // public function __construct()
+    // {
+    //     $this->middleware("auth:sanctum");
+    // }
+
     public function index()
     {
         $students = Student::all();
+        if ($students->isEmpty()) {
+            return response()->json(204);
+        } else {
+            $data = [
+                'message' => 'Get all students',
+                'data' => $students,
+            ];
 
-        $data = [
-            'message' => 'Get all students',
-            'data' => $students,
-        ];
-
-        return response()->json($data, 200);
+            return response()->json($data, 200);
+        }
     }
 
     public function store(Request $request)
     {
+        $request->validate([
+            'nama' => 'required',
+            'nim' => 'required',
+            'email' => 'required|email',
+            'jurusan' => 'required'
+        ]);
+
         $input = [
             'nama' => $request->nama,
             'nim' => $request->nim,
             'email' => $request->email,
-            'jurusan' => $request->jurusan
+            'jurusan' => $request->jurusan,
         ];
 
         $students = Student::create($input);
@@ -41,36 +56,77 @@ class StudentController extends Controller
 
     public function update(Request $request, $id)
     {
-        $input = [
-            'nama' => $request->nama,
-            'nim' => $request->nim,
-            'email' => $request->email,
-            'jurusan' => $request->jurusan
-        ];
-
         $students = Student::find($id);
 
-        $students->save($input);
+        $request->validate([
+            'nama' => 'required',
+            'nim' => 'required',
+            'email' => 'required|email',
+            'jurusan' => 'required'
+        ]);
 
-        $data = [
-            'message' => 'Student is update successfully',
-            'data' => $students,
-        ];
+        if ($students) {
+            $input = [
+                'nama' => $request->nama ?? $students->nama,
+                'nim' => $request->nim ?? $students->nim,
+                'email' => $request->email ?? $students->email,
+                'jurusan' => $request->jurusan ?? $students->jurusan
+            ];
 
-        return response()->json($data, 200);
+            $students->update($input);
+
+            $data = [
+                'message' => 'Student is update successfully',
+                'data' => $students,
+            ];
+
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => 'Student not found',
+            ];
+            return response()->json($data, 404);
+        }
     }
 
     public function destroy($id)
     {
         $students = Student::find($id);
+        if ($students) {
+            $students->delete();
 
-        $students->delete();
+            $data = [
+                'message' => 'Delete students successfully',
+                'data' => $students,
+            ];
 
-        $data = [
-            'message' => 'Delete students successfully',
-            'data' => $students,
-        ];
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => 'Student not found',
+            ];
 
-        return response()->json($data, 203);
+            return response()->json($data, 404);
+        }
+    }
+
+    public function show($id)
+    {
+        $students = Student::find($id);
+
+        if ($students) {
+            $data = [
+                'message' => 'Get detail student',
+                'data' => $students,
+            ];
+
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => 'Student not found',
+            ];
+
+            return response()->json($data, 404);
+        }
     }
 }
